feat(tools): add deferred tag and danger marks to build_tools_section
- Pure deferred group header now shows "(deferred)"
- Tools with is_read_only flag get " (read-only)" suffix
- Tools with is_destructive flag get " (destructive)" suffix
- Eager tools: **Name**(danger): detail
- Deferred tools: Name(danger)
- Extract registry tests to separate file to stay under line limit

Source: issue #653
Type: feat
